+++ DatadogCore/Tests/Datadog/Mocks/RUMDataModelMocks.swift
@@ -256,6 +256,7 @@ extension RUMResourceEvent: RandomMockable {
                 firstByte: .init(duration: .mockRandom(), start: .mockRandom()),
                 id: .mockRandom(),
                 method: .mockRandom(),
+                protocol: nil,
                 provider: .init(
                     domain: .mockRandom(),
                     name: .mockRandom(),
@@ -478,6 +479,7 @@ extension RUMLongTaskEvent: RandomMockable {
                 isFrozenFrame: .mockRandom(),
                 renderStart: nil,
                 scripts: nil,
+                startTime: nil,
                 styleAndLayoutStart: nil
             ),
             os: .mockRandom(),